move the function to create repository link from StudentTable to format
