+++ labtool2.0/src/components/StudentTable.js
@@ -9,6 +9,7 @@ import { associateTeacherToStudent } from '../services/assistant'
 import { showAssistantDropdown, showTagDropdown, selectTeacher, selectTag, selectStudent, unselectStudent, selectAllStudents, unselectAllStudents } from '../reducers/coursePageLogicReducer'
 import { createDropdownTeachers, createDropdownTags } from '../util/dropdown'
 import useLegacyState from '../hooks/legacyState'
+import { createRepositoryLink } from '../util/format'
 
 const { Fragment } = React
 
@@ -279,26 +280,26 @@ export const StudentTable = props => {
     return indents
   }
 
-  const createRepositoryLink = url => {
-    let cleanUrl = url
-    cleanUrl = cleanUrl.replace(/^https?:\/\//, '')
+  // const createRepositoryLink = url => {
+  //   let cleanUrl = url
+  //   cleanUrl = cleanUrl.replace(/^https?:\/\//, '')
 
-    if (cleanUrl.startsWith('github.com/')) {
-      const cleanUrlNoGithub = cleanUrl.substring('github.com/'.length)
-      cleanUrl = (
-        <Fragment>
-          <Icon name="github" color="black" />
-          {cleanUrlNoGithub}
-        </Fragment>
-      )
-    }
+  //   if (cleanUrl.startsWith('github.com/')) {
+  //     const cleanUrlNoGithub = cleanUrl.substring('github.com/'.length)
+  //     cleanUrl = (
+  //       <Fragment>
+  //         <Icon name="github" color="black" />
+  //         {cleanUrlNoGithub}
+  //       </Fragment>
+  //     )
+  //   }
 
-    return (
-      <a href={url} target="_blank" rel="noopener noreferrer">
-        {cleanUrl}
-      </a>
-    )
-  }
+  //   return (
+  //     <a href={url} target="_blank" rel="noopener noreferrer">
+  //       {cleanUrl}
+  //     </a>
+  //   )
+  // }
 
   const createStudentTableRow = (showColumn, data, extraColumns, dropDownTags, dropDownTeachers, { rowClassName, allowReview }) => (
     <Table.Row key={data.id} className={rowClassName}>
